local_api/pyproject.toml
[project]
name = "local_api"
version = "0.1.0"
description = ""
authors = [{ name = "Samuel Stidham", email = "samuel.stidham@snhu.edu" }]
readme = "README.md"
requires-python = ">=3.13, <4.0"
dependencies = [
    "flask (>=3.1.2,<4.0.0)",
    "twilio (>=9.8.4,<10.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "flask-cors (>=6.0.1,<7.0.0)",
]

[tool.poetry]
packages = [{ include = "local_api", from = "src" }]


[tool.poetry.group.dev.dependencies]
black = "^25.9.0"
pytest = "^8.4.2"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
